# Implement CSE for dynamo guards.