@ysiraichi 2023-04-06
In summary, it implements the CSE in 3 steps:

1. Parse each guard expression into an AST
2. Traverse each parsed AST, counting the number of occurrences at each level
    - Holds a map from the stringified node (using `ast.unparse` function) to a counter
3. Traverse each parsed AST, replacing the nodes that occur more than a threshold (currently, 2) by a new variable

**Returns:**
    - the assignments of each new variable to its corresponding expression 
    - the new guard expressions