@ezyang 2023-04-07
This is a very nice improvement, worth 2-4%

<img width="769" alt="image" src="https://user-images.githubusercontent.com/13564/230692267-5eb4355b-7f7f-4bdb-9273-3396fff0b769.png">