@ysiraichi 2023-04-08
So, I think I can make this PR better (I think I can finish the implementation tomorrow) by:

1. We can make it work without `ast.unparse` (eliminating the quadratic time) by using the column information from parsing
2. Use `ALLOW_NODE_TYPES` (the inverse of what we have now) so as to filter the nodes interesting to the CSE, namely: `Subscript`, `Attribute`, and `Call` (for `getattr` calls)
3. Divide the guard function into non-ShapeEnv guards and ShapeEnv guards (which will free us from the try-catch)

@ezyang 
What do you think?