@voznesenskym 2023-04-08
> 3. Divide the guard function into non-ShapeEnv guards and ShapeEnv guards (which will free us from the try-catch)

We have prior art for this already, in terms of treating shape_env code separately from other added guard code.