@ezyang 2023-04-08
> We can make it work without ast.unparse (eliminating the quadratic time) by using the column information from parsing

By using the original string? That's seems brittle. According to the perf run your change doesn't materially change compile time, so I think the quadratic term doesn't matter, leave it as is (with a comment saying that it's quadratic.)

> Use ALLOW_NODE_TYPES (the inverse of what we have now) so as to filter the nodes interesting to the CSE, namely: Subscript, Attribute, and Call (for getattr calls)

Sure, though IMO it's not a big deal.

> Divide the guard function into non-ShapeEnv guards and ShapeEnv guards (which will free us from the try-catch)

As Voz says, this is the minimum you will need to do, but I don't think it is enough. Imagine something like:

```
isinstance(x, list)
len(x) == 2
isinstance(x[0], list)
len(x[0]) == 2
isinstance(x[0][0], int)
isinstance(x[0][1], int)
isinstance(x[1], int)
```

Naive CSE will move x[0] access before we have confirmed that x is a list.