@ezyang 2023-04-09
I think a useful thing to know is what the performance difference between:

```
return a and b and c
```

vs

```
if not a:
  return False
if not b:
  return False
if not c:
  return False
return True
```

is. If there is not much difference, then the easiest thing to do is to stop compiling guards as a big list of `and` statements and then you can CSE at the first point a subexpression is used, without requiring to reorder things.