@ysiraichi 2023-04-09
> According to the perf run your change doesn't materially change compile time, so I think the quadratic term doesn't matter

Not sure what you mean by that. By compile time, do you mean after we introduce JIT to the guards?