@ysiraichi 2023-04-09
Find below the benchmark (torchbench) results for the 2 versions @ezyang mentioned. In summary:

- They don't seem to be significantly different
- There was 1 benchmark that got >1.1x speedup: `resnext50_32x4d`
- There were 2 benchmarks that got <0.9x speedup: `dlrm` (0.88x) and `hf_T5_large` (0.81x)
- There were 3 benchmarks that got <0.98x speedup: `hf_T5`, `hf_T5_base`, and `mobilenet_v2`

<details>
<summary> Raw Results </summary>

- `baseline` refers to the commit `1d7133c542f` (February 15)
- `if-chain` refers to the second version (literally, a chain of ifs)

| Name                           | Latency (baseline) | Latency (if-chain) | Speedup |
|--------------------------------|--------------------|--------------------|---------|
| BERT_pytorch                   | 38.2049            | 38.1404            | 1.0017  |
| Background_Matting             | 148.4077           | 148.1106           | 1.0020  |
| LearningToPaint                | 42.1799            | 42.1211            | 1.0014  |
| Super_SloMo                    | 319.2586           | 319.4182           | 0.9995  |
| alexnet                        | 34.7081            | 34.6411            | 1.0019  |
| dcgan                          | 25.2886            | 25.2632            | 1.0010  |
| densenet121                    | 190.7768           | 192.6175           | 0.9904  |
| dlrm                           | 2.0278             | 2.2835             | 0.8880  |
| drq                            | 0.7249             | 0.7233             | 1.0022  |
| fastNLP_Bert                   | 17.2611            | 17.1888            | 1.0042  |
| functorch_dp_cifar10           | 20.6523            | 20.6971            | 0.9978  |
| functorch_maml_omniglot        | 0.6388             | 0.6464             | 0.9882  |
| hf_Albert                      | 15.0872            | 15.0448            | 1.0028  |
| hf_Bart                        | 18.9939            | 18.8683            | 1.0067  |
| hf_Bert                        | 17.6168            | 17.9100            | 0.9836  |
| hf_DistilBert                  | 8.0332             | 8.0316             | 1.0002  |
| hf_GPT2                        | 37.0473            | 37.0166            | 1.0008  |
| hf_GPT2_large                  | 199.5633           | 199.7220           | 0.9992  |
| hf_Reformer                    | 15.5395            | 15.6129            | 0.9953  |
| hf_T5                          | 90.9850            | 98.2737            | 0.9258  |
| hf_T5_base                     | 305.4032           | 319.5407           | 0.9558  |
| hf_T5_large                    | 127.2385           | 156.5839           | 0.8126  |
| lennard_jones                  | 0.2202             | 0.2198             | 1.0018  |
| maml                           | 114.7211           | 116.2960           | 0.9865  |
| maml_omniglot                  | 0.8444             | 0.8205             | 1.0291  |
| mnasnet1_0                     | 19.5482            | 19.5491            | 1.0000  |
| mobilenet_v2                   | 12.3285            | 12.9194            | 0.9543  |
| mobilenet_v3_large             | 18.2497            | 18.3144            | 0.9965  |
| nvidia_deeprecommender         | 11.4864            | 11.5554            | 0.9940  |
| opacus_cifar10                 | 20.7520            | 20.9510            | 0.9905  |
| pyhpc_equation_of_state        | 6.9967             | 6.9972             | 0.9999  |
| pyhpc_isoneutral_mixing        | 27.9351            | 27.8914            | 1.0016  |
| pyhpc_turbulent_kinetic_energy | 22.3412            | 21.3129            | 1.0482  |
| pytorch_CycleGAN_and_pix2pix   | 14.6733            | 14.6800            | 0.9995  |
| pytorch_stargan                | 51.5640            | 51.6061            | 0.9992  |
| pytorch_unet                   | 113.0233           | 113.1834           | 0.9986  |
| resnet152                      | 172.9542           | 172.9537           | 1.0000  |
| resnet18                       | 7.1986             | 7.1692             | 1.0041  |
| resnet50                       | 66.2987            | 66.3271            | 0.9996  |
| resnext50_32x4d                | 62.5773            | 55.4107            | 1.1293  |
| shufflenet_v2_x1_0             | 33.0737            | 32.9855            | 1.0027  |
| soft_actor_critic              | 0.2517             | 0.2538             | 0.9917  |
| speech_transformer             | 21.2798            | 21.2074            | 1.0034  |
| squeezenet1_1                  | 6.3501             | 6.3379             | 1.0019  |
| tacotron2                      | 1095.2057          | 1095.4023          | 0.9998  |
| timm_efficientnet              | 55.2628            | 55.8831            | 0.9889  |
| timm_regnet                    | 212.7725           | 213.2575           | 0.9977  |
| timm_resnest                   | 43.5706            | 43.5622            | 1.0002  |
| timm_vision_transformer        | 10.7047            | 10.7046            | 1.0000  |
| timm_vision_transformer_large  | 378.2258           | 377.8992           | 1.0009  |
| timm_vovnet                    | 81.9470            | 82.6849            | 0.9911  |
| tts_angular                    | 13.3451            | 13.3499            | 0.9996  |
| vgg16                          | 15.5082            | 15.5191            | 0.9993  |
| yolov3                         | 83.8152            | 84.1460            | 0.9961  |

</details>

-----

After looking at the results, I guess it's best to go with the if-chain version. Let me know your thoughts.

(FYI: I will be off until May 1)