@ysiraichi 2023-05-09
I tried running the inference pass for the huggingface benchmark set locally, but could not reproduce the slow downs. One difference is that I ran this PR (`cse`) against the current base commit of this PR (`8556cf208a3`), instead of using `c6c9258` (the baseline in the dashboard). There seemed to be no difference >1%. Should I check the training pass, too?

Let me know what you think.

**Command:**
```bash
$ python benchmarks/dynamo/huggingface.py --cold-start-latency --device cuda --inference --backend inductor --performance --amp
```

<details>
<summary>HF Performance</summary>

| Name                                    | Latency (cse) | Latency (baseline) | Speedup (cse / baseline) |
|-----------------------------------------|--------------:|-------------------:|-------------------------:|
| AlbertForMaskedLM                       |       97.7787 |            98.0174 |                   0.9976 |
| AlbertForQuestionAnswering              |       97.6842 |            97.5757 |                   1.0011 |
| AllenaiLongformerBase                   |       35.8791 |            35.8827 |                   0.9999 |
| BartForCausalLM                         |       31.4006 |            31.4547 |                   0.9983 |
| BartForConditionalGeneration            |       40.6628 |            40.6341 |                   1.0007 |
| BertForMaskedLM                         |       32.2703 |            32.2552 |                   1.0005 |
| BertForQuestionAnswering                |       25.8519 |            25.7707 |                   1.0032 |
| BlenderbotForCausalLM                   |       22.9920 |            23.0112 |                   0.9992 |
| BlenderbotSmallForCausalLM              |       16.8877 |            16.9482 |                   0.9964 |
| BlenderbotSmallForConditionalGeneration |       32.6727 |            32.5439 |                   1.0040 |
| CamemBert                               |       32.5567 |            32.6716 |                   0.9965 |
| DebertaForMaskedLM                      |       18.2995 |            18.3753 |                   0.9959 |
| DebertaForQuestionAnswering             |       25.4861 |            25.4816 |                   1.0002 |
| DebertaV2ForMaskedLM                    |       22.7355 |            22.8053 |                   0.9969 |
| DebertaV2ForQuestionAnswering           |       18.7498 |            18.7575 |                   0.9996 |
| DistilBertForMaskedLM                   |       32.9714 |            33.0346 |                   0.9981 |
| DistilBertForQuestionAnswering          |       38.8450 |            38.8282 |                   1.0004 |
| DistillGPT2                             |       23.4698 |            23.4863 |                   0.9993 |
| ElectraForCausalLM                      |       19.5771 |            19.5967 |                   0.9990 |
| ElectraForQuestionAnswering             |       25.6321 |            25.6516 |                   0.9992 |
| GPT2ForSequenceClassification           |       17.1804 |            17.1423 |                   1.0022 |
| GoogleFnet                              |       19.3235 |            19.4007 |                   0.9960 |
| LayoutLMForMaskedLM                     |       32.3386 |            32.2846 |                   1.0017 |
| LayoutLMForSequenceClassification       |       25.8167 |            25.8316 |                   0.9994 |
| M2M100ForConditionalGeneration          |       28.7866 |            28.7970 |                   0.9996 |
| MBartForCausalLM                        |       32.0356 |            32.0441 |                   0.9997 |
| MBartForConditionalGeneration           |       41.3213 |            41.3155 |                   1.0001 |
| MT5ForConditionalGeneration             |       12.4830 |            12.4764 |                   1.0005 |
| MegatronBertForCausalLM                 |       24.6044 |            24.5585 |                   1.0019 |
| MegatronBertForQuestionAnswering        |       42.0026 |            41.9453 |                   1.0014 |
| MobileBertForMaskedLM                   |       41.0459 |            41.1578 |                   0.9973 |
| MobileBertForQuestionAnswering          |       27.7888 |            27.6717 |                   1.0042 |
| OPTForCausalLM                          |       27.8872 |            27.8946 |                   0.9997 |
| PLBartForCausalLM                       |       26.8276 |            26.8591 |                   0.9988 |
| PLBartForConditionalGeneration          |       28.5260 |            28.4547 |                   1.0025 |
| PegasusForCausalLM                      |       23.4089 |            23.4320 |                   0.9990 |
| PegasusForConditionalGeneration         |       49.5742 |            49.5978 |                   0.9995 |
| RobertaForCausalLM                      |       32.2950 |            32.3064 |                   0.9996 |
| RobertaForQuestionAnswering             |       25.7564 |            25.7716 |                   0.9994 |
| Speech2Text2ForCausalLM                 |       15.2806 |            15.2656 |                   1.0010 |
| T5ForConditionalGeneration              |       18.4347 |            18.4365 |                   0.9999 |
| T5Small                                 |       18.4758 |            18.4288 |                   1.0026 |
| TrOCRForCausalLM                        |       46.7501 |            46.7689 |                   0.9996 |
| XGLMForCausalLM                         |       15.8072 |            15.8674 |                   0.9962 |
| XLNetLMHeadModel                        |       64.7978 |            64.4825 |                   1.0049 |
| YituTechConvBert                        |       36.4039 |            36.3920 |                   1.0003 |

</details>